Moves Thoughts tests into dedicated directory
