+++ tests/Feature/Thought/CreateThoughtTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature;
+namespace Feature\Thought;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
